+++ baseband_tasks/__init__.py
@@ -1,11 +1,10 @@
 # Licensed under the GPLv3 - see LICENSE
 """Radio baseband data reduction."""
 
-# Packages may add whatever they like to this file, but
-# should keep this content at the top.
-# ----------------------------------------------------------------------------
-from ._astropy_init import *   # noqa
-# ----------------------------------------------------------------------------
+try:
+    from .version import version as __version__
+except ImportError:
+    __version__ = ''
 
 # Define minima for the documentation, but do not bother to explicitly check.
 __minimum_python_version__ = '3.7'