+++ baseband_tasks/__init__.py
@@ -8,5 +8,5 @@ except ImportError:
 
 # Define minima for the documentation, but do not bother to explicitly check.
 __minimum_python_version__ = '3.7'
-__minimum_baseband_version__ = '4.1'
-__minimum_astropy_version__ = '4.0'
+__minimum_baseband_version__ = '4.2'
+__minimum_astropy_version__ = '5.1'