+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.4 (unreleased)
 ================
 
+- Numpy >= 2.0 is now supported.
 - The minimum versions required by baseband-tasks are now python 3.7,
   baseband 4.1, and its minimum dependencies (astropy 4.0 and numpy 1.17).
 - Remove support for the deprecated ``astropy`` test runner. This means it is